Add files via upload
We are creating an e-commerce system which allows customers to
login to the system, perform some shopping operations like purchasing products, viewing
order history and showing user consumption reports. Besides, admin users need to be
created to manage the whole system, who are able to create/delete/view customers,
products and all the orders. Except for the management part, admin users can view the
statistical figures about this system. Since the whole system is executed in the command line
system, it is better to design a well-formatted interface and always show proper messages to
guide users to use your system. I